+++ internal/storage/metrics.go
@@ -5,25 +5,6 @@ import (
 	"fmt"
 )
 
-// MetricType is enum-like type.
-// We are using struct instead of string, to ensure about immutability.
-type MetricType struct {
-	s string
-}
-
-func (u MetricType) IsZero() bool {
-	return u == MetricType{}
-}
-
-func (u MetricType) String() string {
-	return u.s
-}
-
-var (
-	Gauge   = MetricType{"gauge"}
-	Counter = MetricType{"counter"}
-)
-
 func NewMetricTypeString(metricType string) (MetricType, error) {
 	switch metricType {
 	case "gauge":
@@ -34,52 +15,6 @@ func NewMetricTypeString(metricType string) (MetricType, error) {
 
 	return MetricType{}, errors.New(fmt.Sprintf("invalid '%s' MetricType", metricType))
 }
-
-// MetricType is enum-like type.
-// We are using struct instead of string, to ensure about immutability.
-type MetricName struct {
-	s string
-}
-
-func (u MetricName) IsZero() bool {
-	return u == MetricName{}
-}
-
-func (u MetricName) String() string {
-	return u.s
-}
-
-var (
-	Alloc         = MetricName{"alloc"}
-	BuckHashSys   = MetricName{"buckhashsys"}
-	Frees         = MetricName{"frees"}
-	GCSys         = MetricName{"gcsys"}
-	HeapAlloc     = MetricName{"heapalloc"}
-	HeapIdle      = MetricName{"heapidle"}
-	HeapInuse     = MetricName{"heapinuse"}
-	HeapObjects   = MetricName{"heapobjects"}
-	HeapReleased  = MetricName{"heapreleased"}
-	HeapSys       = MetricName{"heapsys"}
-	LastGC        = MetricName{"lastgc"}
-	Lookups       = MetricName{"lookups"}
-	MCacheInuse   = MetricName{"mcacheinuse"}
-	MCacheSys     = MetricName{"mcachesys"}
-	MSpanInuse    = MetricName{"mspaninuse"}
-	MSpanSys      = MetricName{"mspansys"}
-	Mallocs       = MetricName{"mallocs"}
-	NextGC        = MetricName{"nextgc"}
-	OtherSys      = MetricName{"othersys"}
-	PauseTotalNs  = MetricName{"pausetotalns"}
-	StackInuse    = MetricName{"stackinuse"}
-	StackSys      = MetricName{"stacksys"}
-	RandomValue   = MetricName{"randomvalue"}
-	Sys           = MetricName{"sys"}
-	NumGC         = MetricName{"numgc"}
-	NumForcedGC   = MetricName{"numforcedgc"}
-	GCCPUFraction = MetricName{"gccpufraction"}
-	NumGoroutine  = MetricName{"numgoroutine"}
-	PollCount     = MetricName{"pollcount"}
-)
 
 func NewMetricNameString(metricName string) (MetricName, error) {
 	switch metricName {
@@ -144,13 +79,6 @@ func NewMetricNameString(metricName string) (MetricName, error) {
 	}
 
 	return MetricName{}, errors.New(fmt.Sprintf("invalid '%s' MetricName", metricName))
-}
-
-type Metric struct {
-	metricType MetricType
-	metricName MetricName
-	counter    int
-	value      string
 }
 
 func (u *Metric) NewMetricString(metricName string, metricType string, value string) error {
